feat: 将集数统计返回的百分比换成集数数量，修复sqlite的类型错误

+++ orm/entity/src/anime.rs
@@ -11,7 +11,7 @@ pub struct Model {
     pub anime_info: serde_json::Value,
     pub is_search: bool,
     pub is_lock: bool,
-    pub progress: u8,
+    pub progress: u32,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
